Moved to STXXL 1.4.1. Created unordered map wrapper.

+++ include/stxxl/bits/common/binary_buffer.h
@@ -30,6 +30,9 @@
 
 STXXL_BEGIN_NAMESPACE
 
+//! \addtogroup support
+//! \{
+
 /*!
  * binary_buffer represents a dynamically growable area of memory, which can be
  * modified by appending integral data types via put() and other basic
@@ -278,29 +281,29 @@ public:
     inline binary_buffer & put_varint(uint32 v)
     {
         if (v < 128) {
-            put<uint8>(v);
+            put<uint8>(uint8(v));
         }
         else if (v < 128 * 128) {
-            put<uint8>(((v >> 0) & 0x7F) | 0x80);
-            put<uint8>(((v >> 7) & 0x7F));
+            put<uint8>((uint8)(((v >> 0) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 7) & 0x7F));
         }
         else if (v < 128 * 128 * 128) {
-            put<uint8>(((v >> 0) & 0x7F) | 0x80);
-            put<uint8>(((v >> 7) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F));
+            put<uint8>((uint8)(((v >> 0) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 7) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 14) & 0x7F));
         }
         else if (v < 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 0) & 0x7F) | 0x80);
-            put<uint8>(((v >> 7) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F));
+            put<uint8>((uint8)(((v >> 0) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 7) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 21) & 0x7F));
         }
         else {
-            put<uint8>(((v >> 0) & 0x7F) | 0x80);
-            put<uint8>(((v >> 7) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F));
+            put<uint8>((uint8)(((v >> 0) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 7) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 28) & 0x7F));
         }
 
         return *this;
@@ -316,79 +319,79 @@ public:
     inline binary_buffer & put_varint(uint64 v)
     {
         if (v < 128) {
-            put<uint8>(v);
+            put<uint8>(uint8(v));
         }
         else if (v < 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F));
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 07) & 0x7F));
         }
         else if (v < 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F));
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 14) & 0x7F));
         }
         else if (v < 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F));
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 21) & 0x7F));
         }
-        else if (v < 128ULL * 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F));
+        else if (v < ((uint64)128) * 128 * 128 * 128 * 128) {
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 28) & 0x7F));
         }
-        else if (v < 128ULL * 128 * 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F) | 0x80);
-            put<uint8>(((v >> 35) & 0x7F));
+        else if (v < ((uint64)128) * 128 * 128 * 128 * 128 * 128) {
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 28) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 35) & 0x7F));
         }
-        else if (v < 128ULL * 128 * 128 * 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F) | 0x80);
-            put<uint8>(((v >> 35) & 0x7F) | 0x80);
-            put<uint8>(((v >> 42) & 0x7F));
+        else if (v < ((uint64)128) * 128 * 128 * 128 * 128 * 128 * 128) {
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 28) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 35) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 42) & 0x7F));
         }
-        else if (v < 128ULL * 128 * 128 * 128 * 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F) | 0x80);
-            put<uint8>(((v >> 35) & 0x7F) | 0x80);
-            put<uint8>(((v >> 42) & 0x7F) | 0x80);
-            put<uint8>(((v >> 49) & 0x7F));
+        else if (v < ((uint64)128) * 128 * 128 * 128 * 128 * 128 * 128 * 128) {
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 28) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 35) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 42) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 49) & 0x7F));
         }
-        else if (v < 128ULL * 128 * 128 * 128 * 128 * 128 * 128 * 128 * 128) {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F) | 0x80);
-            put<uint8>(((v >> 35) & 0x7F) | 0x80);
-            put<uint8>(((v >> 42) & 0x7F) | 0x80);
-            put<uint8>(((v >> 49) & 0x7F) | 0x80);
-            put<uint8>(((v >> 56) & 0x7F));
+        else if (v < ((uint64)128) * 128 * 128 * 128 * 128 * 128 * 128 * 128 * 128) {
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 28) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 35) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 42) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 49) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 56) & 0x7F));
         }
         else {
-            put<uint8>(((v >> 00) & 0x7F) | 0x80);
-            put<uint8>(((v >> 07) & 0x7F) | 0x80);
-            put<uint8>(((v >> 14) & 0x7F) | 0x80);
-            put<uint8>(((v >> 21) & 0x7F) | 0x80);
-            put<uint8>(((v >> 28) & 0x7F) | 0x80);
-            put<uint8>(((v >> 35) & 0x7F) | 0x80);
-            put<uint8>(((v >> 42) & 0x7F) | 0x80);
-            put<uint8>(((v >> 49) & 0x7F) | 0x80);
-            put<uint8>(((v >> 56) & 0x7F) | 0x80);
-            put<uint8>(((v >> 63) & 0x7F));
+            put<uint8>((uint8)(((v >> 00) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 07) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 14) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 21) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 28) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 35) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 42) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 49) & 0x7F) | 0x80));
+            put<uint8>((uint8)(((v >> 56) & 0x7F) | 0x80));
+            put<uint8>((uint8)((v >> 63) & 0x7F));
         }
 
         return *this;
@@ -639,6 +642,8 @@ public:
         return br;
     }
 };
+
+//! \}
 
 STXXL_END_NAMESPACE
